Extracted IBaseRepository to help with injection in consuming projects.
