+++ DataFace/DataFace.csproj
@@ -51,6 +51,7 @@
     <Compile Include="Core\Column.cs" />
     <Compile Include="Core\DataFaceException.cs" />
     <Compile Include="Core\CommandOptions.cs" />
+    <Compile Include="Core\IBaseRepository.cs" />
     <Compile Include="Core\IDatabaseConnection.cs" />
     <Compile Include="Core\IEnumerableExtensions.cs" />
     <Compile Include="Core\ICommand.cs" />
@@ -61,6 +62,7 @@
     <Compile Include="Core\ResultSetConverter.cs" />
     <Compile Include="Core\ResultSet.cs" />
     <Compile Include="Core\Row.cs" />
+    <Compile Include="Core\TransactionContext.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Core\ResultSetAttribute.cs" />
     <Compile Include="Core\ResultSetType.cs" />
